python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openlvm"
version = "0.1.0"
description = "Performance-first, agent-native virtual machine with built-in testing and chaos simulation."
readme = "../README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "diiviikk5", email = "diiviikk5@users.noreply.github.com" },
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
eval = [
    "deepeval>=0.21.0",
]
tracing = [
    "traceloop-sdk>=0.15.0",
]
mcp = [
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
openlvm = "openlvm.cli:app"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --import-mode=importlib"
testpaths = [
    "tests",
]
